j3

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ladawi <ladawi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 13:37:57 by ladawi            #+#    #+#             */
-/*   Updated: 2020/01/25 16:39:55 by ladawi           ###   ########.fr       */
+/*   Updated: 2021/05/26 08:41:30 by ladawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ char			*ft_strnstr(const char *s1, const char *s2, size_t len);
 int				ft_strncmp(const char *s1, const char *s2, size_t n);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
 size_t			ft_strlcat(char *dest, const char *src, size_t size);
-int				ft_atoi(const char *nptr);
+long int				ft_atoi(const char *nptr);
 void			*ft_calloc(size_t nmemb, size_t size);
 char			*ft_strdup(const char *s);
 char			*ft_strjoin(char const *s1, char *s2);
