+++ libft/src/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: ladawi <ladawi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 14:01:17 by ladawi            #+#    #+#             */
-/*   Updated: 2021/05/24 13:42:46 by ladawi           ###   ########.fr       */
+/*   Updated: 2021/05/26 08:41:17 by ladawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 #include <stdio.h>
 
-int		ft_atoi(const char *str)
+long int		ft_atoi(const char *str)
 {
-	int		nb;
+	long int		nb;
 	int		sign;
 
 	nb = 0;
